requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
backcall==0.2.0
black==21.12b0
boto==2.49.0
boto3==1.20.46
botocore==1.23.46
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
decorator==5.1.1
executing==0.8.2
frozenlist==1.3.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipdb==0.13.9
ipython==8.0.1
itemadapter==0.4.0
itemloaders==1.0.4
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jsonpointer==2.2
jsonschema==3.2.0
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
multidict==6.0.2
mypy-extensions==0.4.3
parsel==1.6.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
priority==1.3.0
prompt-toolkit==3.0.24
Protego==0.1.16
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyDispatcher==2.0.5
Pygments==2.11.2
pyOpenSSL==21.0.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-slugify==5.0.2
queuelib==1.6.2
requests==2.27.1
rfc3987==1.3.8
s3transfer==0.5.0
schematics==2.1.1
Scrapy==2.5.1
scrapyd==1.3.0
scrapyd-client==1.2.0
service-identity==21.1.0
six==1.16.0
slackclient==2.9.3
spidermon==1.16.2
stack-data==0.1.4
strict-rfc3339==0.7
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
traitlets==5.1.1
Twisted==21.7.0
typing_extensions==4.0.1
urllib3==1.26.8
w3lib==1.22.0
wcwidth==0.2.5
webcolors==1.11.1
yarl==1.7.2
zope.interface==5.4.0
